refactor(dashboard): extract deployment log viewer into its own component
