+++ dashboard/src/routes/deployments/[id]/+page.svelte
@@ -3,22 +3,19 @@
   import { goto } from '$app/navigation';
   import { page } from '$app/stores';
   import {
-    fetchLogsSnapshot,
     getDeployment,
     getDeploymentHealthChecks,
     getDeploymentMetrics,
     listDeploymentEvents,
-    streamLogs,
     type DeploymentDetail,
     type DeploymentEvent,
     type DeploymentStats,
     type EnvValue,
     type HealthCheck,
-    type HealthCheckResult,
-    type LogEntry,
-    type LogStreamHandle
+    type HealthCheckResult
   } from '$lib/api';
   import CopyButton from '$lib/CopyButton.svelte';
+  import DeploymentLogs from '$lib/DeploymentLogs.svelte';
   import { getToken } from '$lib/auth';
   import { formatBytes, formatDate, timeAgo } from '$lib/utils';
 
@@ -33,130 +30,6 @@
   let poll: ReturnType<typeof setInterval> | null = null;
 
   let id = $derived($page.params.id ?? '');
-
-  /** Hard cap to bound memory when a chatty deployment streams for hours.
-   *  Older entries fall off the top once we cross this. */
-  const LOG_RING_MAX = 2000;
-
-  let logs = $state<LogEntry[]>([]);
-  let logsLoading = $state(false);
-  let logsError = $state<string | null>(null);
-  let logsFollow = $state(true);
-  let logsTail = $state(200);
-  let logsAutoScroll = $state(true);
-  let logsContainerFilter = $state('');
-  let logsStream: LogStreamHandle | null = null;
-  let logsEl: HTMLDivElement | null = $state(null);
-
-  function pushLog(entry: LogEntry) {
-    logs.push(entry);
-    if (logs.length > LOG_RING_MAX) {
-      logs.splice(0, logs.length - LOG_RING_MAX);
-    }
-  }
-
-  function stopStream() {
-    if (logsStream) {
-      logsStream.close();
-      logsStream = null;
-    }
-  }
-
-  async function loadSnapshot() {
-    if (!id) {
-      return;
-    }
-    logsLoading = true;
-    logsError = null;
-    try {
-      const snap = await fetchLogsSnapshot(id, {
-        tail: logsTail,
-        container: logsContainerFilter || undefined
-      });
-      logs = snap;
-    } catch (e) {
-      logsError = e instanceof Error ? e.message : String(e);
-    } finally {
-      logsLoading = false;
-    }
-  }
-
-  async function startStream() {
-    if (!id) {
-      return;
-    }
-    stopStream();
-    logsError = null;
-    try {
-      logsStream = await streamLogs(
-        id,
-        { tail: logsTail, container: logsContainerFilter || undefined },
-        (entry) => {
-          pushLog(entry);
-        },
-        () => {
-          // EventSource fires a generic Event on disconnect; surface a hint
-          // without killing the page. The browser will auto-reconnect using
-          // the same URL while the ticket is still within its TTL.
-          logsError = 'stream interrupted, reconnecting…';
-        }
-      );
-    } catch (e) {
-      logsError = e instanceof Error ? e.message : String(e);
-    }
-  }
-
-  async function refreshLogs() {
-    stopStream();
-    // Don't clear `logs` here: emptying the array collapses the viewport,
-    // which makes the page jump to the top because the scroll position
-    // becomes out of range. Keep the old lines until the snapshot lands
-    // and then swap atomically inside loadSnapshot().
-    await loadSnapshot();
-    if (logsFollow) {
-      await startStream();
-    }
-  }
-
-  async function toggleFollow() {
-    logsFollow = !logsFollow;
-    if (logsFollow) {
-      await startStream();
-    } else {
-      stopStream();
-    }
-  }
-
-  function levelClass(level: string): string {
-    const l = level.toLowerCase();
-    if (l === 'error' || l === 'err' || l === 'critical' || l === 'fatal') {
-      return 'log-error';
-    }
-    if (l === 'warn' || l === 'warning') {
-      return 'log-warn';
-    }
-    if (l === 'info' || l === 'notice') {
-      return 'log-info';
-    }
-    if (l === 'debug') {
-      return 'log-debug';
-    }
-    return 'log-unknown';
-  }
-
-  $effect(() => {
-    if (logsAutoScroll && logsEl && logs.length > 0) {
-      // Read .length to register the dependency, then defer the scroll
-      // until after the DOM has the new row.
-      const _ = logs.length;
-      void _;
-      queueMicrotask(() => {
-        if (logsEl) {
-          logsEl.scrollTop = logsEl.scrollHeight;
-        }
-      });
-    }
-  });
 
   async function refresh() {
     if (!id) {
@@ -212,7 +85,6 @@
       return;
     }
     void refresh();
-    void refreshLogs();
     poll = setInterval(() => void refresh(), 5000);
   });
 
@@ -220,7 +92,6 @@
     if (poll) {
       clearInterval(poll);
     }
-    stopStream();
   });
 
   function statusKind(s: string): 'success' | 'warn' | 'danger' | 'neutral' {
@@ -683,58 +554,7 @@
     </section>
   {/if}
 
-  <section class="card">
-    <header class="section-head logs-head">
-      <h2>Logs</h2>
-      <div class="logs-controls">
-        <label class="logs-control">
-          <span>Tail</span>
-          <select bind:value={logsTail} onchange={refreshLogs}>
-            <option value={100}>100</option>
-            <option value={200}>200</option>
-            <option value={500}>500</option>
-            <option value={1000}>1000</option>
-          </select>
-        </label>
-        <label class="logs-control checkbox">
-          <input type="checkbox" bind:checked={logsAutoScroll} />
-          <span>Auto-scroll</span>
-        </label>
-        <button
-          class="btn-secondary small"
-          class:active={logsFollow}
-          onclick={toggleFollow}
-          title={logsFollow ? 'Pause live stream' : 'Resume live stream'}
-        >
-          {logsFollow ? '⏸ Pause' : '▶ Live'}
-        </button>
-        <button class="btn-secondary small" onclick={refreshLogs} disabled={logsLoading}>
-          Reload
-        </button>
-      </div>
-    </header>
-    {#if logsError}
-      <div class="logs-error">{logsError}</div>
-    {/if}
-    <div class="logs-viewport mono" bind:this={logsEl}>
-      {#if logsLoading && logs.length === 0}
-        <p class="muted pad">Loading logs…</p>
-      {:else if logs.length === 0}
-        <p class="muted pad">No logs yet.</p>
-      {:else}
-        {#each logs as entry, i (i)}
-          <div class="log-line">
-            {#if entry.timestamp}
-              <span class="log-ts">{entry.timestamp}</span>
-            {/if}
-            <span class="log-level {levelClass(entry.level)}">{entry.level}</span>
-            <span class="log-instance">{entry.instance}</span>
-            <span class="log-msg">{entry.message}</span>
-          </div>
-        {/each}
-      {/if}
-    </div>
-  </section>
+  <DeploymentLogs {id} />
 
   {#if events.length > 0}
     <section class="card">
@@ -1079,104 +899,4 @@
     font-size: 0.7rem;
   }
 
-  .logs-head {
-    gap: 0.75rem;
-  }
-  .logs-controls {
-    display: flex;
-    align-items: center;
-    gap: 0.6rem;
-  }
-  .logs-control {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.35rem;
-    font-size: 0.75rem;
-    color: var(--fg-2);
-  }
-  .logs-control select {
-    font-size: 0.78rem;
-    padding: 0.18rem 0.35rem;
-    background: var(--bg-0);
-    color: var(--fg-0);
-    border: 1px solid var(--border);
-    border-radius: var(--radius-sm);
-  }
-  .logs-control.checkbox input {
-    margin: 0;
-  }
-  .btn-secondary.small {
-    font-size: 0.75rem;
-    padding: 0.25rem 0.6rem;
-  }
-  .btn-secondary.active {
-    color: var(--accent);
-    border-color: var(--accent);
-  }
-  .logs-error {
-    padding: 0.45rem 1.125rem;
-    background: var(--warning-bg);
-    color: var(--warning);
-    font-size: 0.78rem;
-    border-bottom: 1px solid var(--border);
-  }
-  .logs-viewport {
-    max-height: 28rem;
-    overflow-y: auto;
-    background: var(--bg-0);
-    font-size: 0.78rem;
-    line-height: 1.45;
-  }
-  .log-line {
-    display: grid;
-    grid-template-columns: auto auto auto 1fr;
-    column-gap: 0.65rem;
-    padding: 0.18rem 1.125rem;
-    border-bottom: 1px solid transparent;
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-  .log-line:hover {
-    background: var(--bg-1);
-  }
-  .log-ts {
-    color: var(--fg-3);
-    font-size: 0.72rem;
-  }
-  .log-level {
-    text-transform: uppercase;
-    font-size: 0.68rem;
-    letter-spacing: 0.04em;
-    padding: 0 0.3rem;
-    border-radius: var(--radius-sm);
-    align-self: center;
-    min-width: 4rem;
-    text-align: center;
-  }
-  .log-error {
-    color: var(--danger);
-    background: var(--danger-bg);
-  }
-  .log-warn {
-    color: var(--warning);
-    background: var(--warning-bg);
-  }
-  .log-info {
-    color: var(--success);
-    background: var(--success-bg);
-  }
-  .log-debug {
-    color: var(--fg-2);
-    background: var(--bg-2);
-  }
-  .log-unknown {
-    color: var(--fg-3);
-  }
-  .log-instance {
-    color: var(--fg-2);
-    font-size: 0.72rem;
-  }
-  .log-msg {
-    color: var(--fg-0);
-  }
 </style>
